fix(round4): 考试安全 + 讨论并发 + 成绩/视频 N+1 — 第四轮盲扫
Phase 27-29 发现 3 P0 + 8 P1 + 6 P2,本次修复:
P0 (3):
- EXAM-NEW-1: QuestionController GET /api/questions 泄露答案
  收紧 @PreAuthorize 至 TEACHER/ADMIN/ACADEMIC(学生端不直接调此接口)
- DISC-NEW-1: DiscussionAdminController.delete() 传 null userId
  导致管理员永远无法删除帖子,改为 SecurityUtil.getCurrentUserId()
- DISC-NEW-2: DiscussionCommentService.like() 读-改-写并发
  改为原子 SQL setSql("like_count = COALESCE(like_count,0)+1")

P1 (3):
- MISC-NEW-1: grades 表缺 UNIQUE → V44 加部分唯一索引
- MISC-NEW-2: VideoServiceImpl.page() chapter N+1
  批量预加载 chapterMap,避免分页场景下 selectById
- EXAM 预存: QuestionService/Controller courseId Long/Integer 类型不一致修复

客户体验影响:
- 学生无法从题库接口获取试题答案(防作弊第一道防线)
- 管理员可正常删除不当讨论(内容管控恢复)
- 评论点赞不再丢失更新
- 成绩不可重复产生
- 视频分页加载减少 1 次 SQL/行

验证:
- mvn compile: SUCCESS
- npm build: SUCCESS
- 31/31 集成测试 PASS

+++ micro-course-api/src/main/java/com/microcourse/service/QuestionService.java
@@ -15,7 +15,7 @@ public interface QuestionService {
 
     void delete(Long id);
 
-    PageResult<QuestionVO> page(Integer courseId, String questionType, Integer difficulty, Integer page, Integer size);
+    PageResult<QuestionVO> page(Long courseId, String questionType, Integer difficulty, Integer page, Integer size);
 
     QuestionVO getById(Long id);
